7 tasks to be done
